script部分をファイルに分割
